1.6v

+++ backend/clothing/models.py
@@ -6,5 +6,5 @@ class cloth(models.Model):
     id = models.AutoField(primary_key=True)
     model = models.CharField(max_length=200)
     size = models.CharField(max_length=4)
-    value = models.IntegerField(blank=True, null=True)
-    amount = models.FloatField(blank=True, null=True)
+    value = models.FloatField(blank=True, null=True)
+    amount = models.IntegerField(blank=True, null=True)
